cicd/requirements.txt
dagger.io==0.2.1
docker==6.0.1
click